+++ pkg/kapis/logging/v1alpha2/register.go
