+++ pkg/kapis/monitoring/v1alpha2/register.go
