+++ pkg/kapis/openpitrix/v1/register.go
