+++ pkg/kapis/servicemesh/metrics/install/install.go
@@ -1,0 +1,16 @@
+package install
+
+import (
+	"github.com/emicklei/go-restful"
+	urlruntime "k8s.io/apimachinery/pkg/util/runtime"
+	"kubesphere.io/kubesphere/pkg/apiserver/runtime"
+	v1alpha22 "kubesphere.io/kubesphere/pkg/kapis/servicemesh/metrics/v1alpha2"
+)
+
+func init() {
+	Install(runtime.Container)
+}
+
+func Install(c *restful.Container) {
+	urlruntime.Must(v1alpha22.AddToContainer(c))
+}
