+++ pkg/kapis/servicemesh/metrics/v1alpha2/swagger-doc.go
