+++ pkg/kapis/tenant/v1alpha2/register.go
