+++ pkg/kapis/terminal/v1alpha2/register.go
