+++ tools/cmd/doc-gen/main.go
@@ -28,18 +28,18 @@ import (
 	"io/ioutil"
 	"kubesphere.io/kubesphere/pkg/apiserver/runtime"
 	"kubesphere.io/kubesphere/pkg/constants"
+	_ "kubesphere.io/kubesphere/pkg/kapis/servicemesh/metrics/install"
 	"log"
 	// Install apis
-	_ "kubesphere.io/kubesphere/pkg/apis/devops/install"
-	_ "kubesphere.io/kubesphere/pkg/apis/iam/install"
-	_ "kubesphere.io/kubesphere/pkg/apis/logging/install"
-	_ "kubesphere.io/kubesphere/pkg/apis/monitoring/install"
-	_ "kubesphere.io/kubesphere/pkg/apis/openpitrix/install"
-	_ "kubesphere.io/kubesphere/pkg/apis/operations/install"
-	_ "kubesphere.io/kubesphere/pkg/apis/resources/install"
-	_ "kubesphere.io/kubesphere/pkg/apis/servicemesh/metrics/install"
-	_ "kubesphere.io/kubesphere/pkg/apis/tenant/install"
-	_ "kubesphere.io/kubesphere/pkg/apis/terminal/install"
+	_ "kubesphere.io/kubesphere/pkg/kapis/devops/install"
+	_ "kubesphere.io/kubesphere/pkg/kapis/iam/install"
+	_ "kubesphere.io/kubesphere/pkg/kapis/logging/install"
+	_ "kubesphere.io/kubesphere/pkg/kapis/monitoring/install"
+	_ "kubesphere.io/kubesphere/pkg/kapis/openpitrix/install"
+	_ "kubesphere.io/kubesphere/pkg/kapis/operations/install"
+	_ "kubesphere.io/kubesphere/pkg/kapis/resources/install"
+	_ "kubesphere.io/kubesphere/pkg/kapis/tenant/install"
+	_ "kubesphere.io/kubesphere/pkg/kapis/terminal/install"
 )
 
 var output string
